switch fix and posible overloads in tests were fixed

+++ tests/cpp/common/test_hist_util.cc
@@ -349,9 +349,9 @@ TEST(hist_util, SparseCutsExternalMemory) {
 }
 
 TEST(hist_util, IndexBinBound) {
-  uint64_t bin_sizes[] = {std::numeric_limits<uint8_t>::max() + 1,
-                          std::numeric_limits<uint16_t>::max() + 1,
-                          std::numeric_limits<uint16_t>::max() + 2};
+  uint64_t bin_sizes[] = {static_cast<uint64_t>(std::numeric_limits<uint8_t>::max()) + 1,
+                          static_cast<uint64_t>(std::numeric_limits<uint16_t>::max()) + 1,
+                          static_cast<uint64_t>(std::numeric_limits<uint16_t>::max()) + 2};
   BinBounds expected_bin_bounds[] = {UINT8_BINS_TYPE,
                                      UINT16_BINS_TYPE,
                                      UINT32_BINS_TYPE};
@@ -371,9 +371,9 @@ TEST(hist_util, IndexBinBound) {
 }
 
 TEST(hist_util, SparseIndexBinBound) {
-  uint64_t bin_sizes[] = {std::numeric_limits<uint8_t>::max() + 1,
-                          std::numeric_limits<uint16_t>::max() + 1,
-                          std::numeric_limits<uint16_t>::max() + 2};
+  uint64_t bin_sizes[] = {static_cast<uint64_t>(std::numeric_limits<uint8_t>::max()) + 1,
+                          static_cast<uint64_t>(std::numeric_limits<uint16_t>::max()) + 1,
+                          static_cast<uint64_t>(std::numeric_limits<uint16_t>::max()) + 2};
   BinBounds expected_bin_bounds[] = {UINT32_BINS_TYPE,
                                      UINT32_BINS_TYPE,
                                      UINT32_BINS_TYPE};
@@ -400,9 +400,9 @@ void CheckIndexData(T* data_ptr, uint32_t* offsets,
 }
 
 TEST(hist_util, IndexBinData) {
-  uint64_t constexpr kBinSizes[] = {std::numeric_limits<uint8_t>::max() + 1,
-                                    std::numeric_limits<uint16_t>::max() + 1,
-                                    std::numeric_limits<uint16_t>::max() + 2};
+  uint64_t constexpr kBinSizes[] = {static_cast<uint64_t>(std::numeric_limits<uint8_t>::max()) + 1,
+                                    static_cast<uint64_t>(std::numeric_limits<uint16_t>::max()) + 1,
+                                    static_cast<uint64_t>(std::numeric_limits<uint16_t>::max()) + 2};
   size_t constexpr kRows = 100;
   size_t constexpr kCols = 10;
 
@@ -433,9 +433,9 @@ TEST(hist_util, IndexBinData) {
 }
 
 TEST(hist_util, SparseIndexBinData) {
-  uint64_t bin_sizes[] = {std::numeric_limits<uint8_t>::max() + 1,
-                          std::numeric_limits<uint16_t>::max() + 1,
-                          std::numeric_limits<uint16_t>::max() + 2};
+  uint64_t bin_sizes[] = {static_cast<uint64_t>(std::numeric_limits<uint8_t>::max()) + 1,
+                          static_cast<uint64_t>(std::numeric_limits<uint16_t>::max()) + 1,
+                          static_cast<uint64_t>(std::numeric_limits<uint16_t>::max()) + 2};
   size_t constexpr kRows = 100;
   size_t constexpr kCols = 10;
